Savepoint. Think the DB sync with access_token is fixed
I still need to make deeper tests to know if the system is stable enough. However, so far I've seen it's not needed to remove the refresh_token in the user's Google account.
Also, regardless of having the user in the local database, it may be created now on remote. And if he exists in both sides, the tokens are successfully sync and updated.